+++ gradoop-flink/src/main/java/org/gradoop/flink/datagen/foodbroker/FoodBroker.java
@@ -31,7 +31,14 @@ import org.gradoop.common.model.impl.pojo.Vertex;
 import org.gradoop.flink.datagen.foodbroker.config.Constants;
 import org.gradoop.flink.datagen.foodbroker.functions.*;
 import org.gradoop.flink.datagen.foodbroker.config.FoodBrokerConfig;
-import org.gradoop.flink.datagen.foodbroker.masterdata.*;
+import org.gradoop.flink.datagen.foodbroker.functions.masterdata.*;
+import org.gradoop.flink.datagen.foodbroker.functions.process.Brokerage;
+import org.gradoop.flink.datagen.foodbroker.functions.process.ComplaintHandling;
+import org.gradoop.flink.datagen.foodbroker.generators.CustomerGenerator;
+import org.gradoop.flink.datagen.foodbroker.generators.EmployeeGenerator;
+import org.gradoop.flink.datagen.foodbroker.generators.LogisticsGenerator;
+import org.gradoop.flink.datagen.foodbroker.generators.ProductGenerator;
+import org.gradoop.flink.datagen.foodbroker.generators.VendorGenerator;
 import org.gradoop.flink.model.impl.GraphCollection;
 import org.gradoop.flink.model.impl.functions.epgm.GraphTransactionTriple;
 import org.gradoop.flink.model.impl.functions.epgm.TransactionEdges;
