+++ gradoop-flink/src/main/java/org/gradoop/flink/datagen/foodbroker/functions/masterdata/Logistics.java
@@ -14,8 +14,8 @@
  * You should have received a copy of the GNU General Public License
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
-
 package org.gradoop.flink.datagen.foodbroker.functions.masterdata;
+
 
 import org.apache.flink.api.common.functions.RichMapFunction;
 import org.apache.flink.configuration.Configuration;
@@ -28,53 +28,20 @@ import org.gradoop.flink.datagen.foodbroker.tuples.MasterDataSeed;
 import java.util.List;
 import java.util.Random;
 
-/**
- * Creates a logistic vertex.
- */
 public class Logistics
   extends RichMapFunction<MasterDataSeed, Vertex> {
-  /**
-   * Class name of the vertex.
-   */
+
   public static final String CLASS_NAME = "Logistics";
-  /**
-   * Acronym for logistics.
-   */
   private static final String ACRONYM = "LOG";
-  /**
-   * List of possible adjectives.
-   */
   private List<String> adjectives;
-  /**
-   * List of possible nouns.
-   */
   private List<String> nouns;
-  /**
-   * List of possible cities.
-   */
   private List<String> cities;
-  /**
-   * Amount of possible adjectives.
-   */
   private Integer adjectiveCount;
-  /**
-   * Amount of possible nouns.
-   */
   private Integer nounCount;
-  /**
-   * Amount of pissible cities.
-   */
   private Integer cityCount;
-  /**
-   * EPGM vertex factory.
-   */
+
   private final VertexFactory vertexFactory;
 
-  /**
-   * Valued constructor.
-   *
-   * @param vertexFactory EPGM vertex factory
-   */
   public Logistics(VertexFactory vertexFactory) {
     this.vertexFactory = vertexFactory;
   }
@@ -83,14 +50,14 @@ public class Logistics
   @Override
   public void open(Configuration parameters) throws Exception {
     super.open(parameters);
-    //load broadcast lists
+
     adjectives = getRuntimeContext()
       .getBroadcastVariable(LogisticsGenerator.ADJECTIVES_BC);
     nouns = getRuntimeContext()
       .getBroadcastVariable(LogisticsGenerator.NOUNS_BC);
     cities = getRuntimeContext()
       .getBroadcastVariable(LogisticsGenerator.CITIES_BC);
-    //get their sizes
+
     nounCount = nouns.size();
     adjectiveCount = adjectives.size();
     cityCount = cities.size();
@@ -98,14 +65,15 @@ public class Logistics
 
   @Override
   public Vertex map(MasterDataSeed seed) throws  Exception {
-    //create standard properties from acronym and seed
-    PropertyList properties = MasterData.createDefaultProperties(seed, ACRONYM);
+    PropertyList properties = MasterData.createDefaultProperties(ACRONYM, seed);
+
     Random random = new Random();
-    //set rnd city and name
+
     properties.set("city", cities.get(random.nextInt(cityCount)));
     properties.set("name",
       adjectives.get(random.nextInt(adjectiveCount)) +
       " " + nouns.get(random.nextInt(nounCount)));
+
     return vertexFactory.createVertex(Logistics.CLASS_NAME, properties);
   }
 }
