+++ gradoop-flink/src/main/java/org/gradoop/flink/datagen/foodbroker/functions/masterdata/MasterDataQualityMapper.java
@@ -24,8 +24,7 @@ import org.gradoop.common.model.impl.pojo.Vertex;
 import org.gradoop.flink.datagen.foodbroker.config.Constants;
 
 /**
- * Creates a master data tuple from the given vertex. The tuple consists of the gradoop id and
- * the quality.
+ * Creates a master data tuple from the given vertex.
  */
 public class MasterDataQualityMapper
   implements MapFunction<Vertex, Tuple2<GradoopId, Float>> {
