+++ gradoop-flink/src/main/java/org/gradoop/flink/datagen/foodbroker/functions/masterdata/ProductPriceMapper.java
@@ -26,8 +26,7 @@ import org.gradoop.flink.datagen.foodbroker.config.Constants;
 import java.math.BigDecimal;
 
 /**
- * Creates a product price tuple from the given vertex. The tuple consists of the gradoop id and
- * the price.
+ * Creates a product price tuple from the given vertex.
  */
 public class ProductPriceMapper implements
   MapFunction<Vertex, Tuple2<GradoopId, BigDecimal>> {
