+++ gradoop-flink/src/main/java/org/gradoop/flink/datagen/foodbroker/functions/masterdata/UserClients.java
@@ -1,20 +1,3 @@
-/*
- * This file is part of Gradoop.
- *
- * Gradoop is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Gradoop is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
- */
-
 package org.gradoop.flink.datagen.foodbroker.functions.masterdata;
 
 import org.apache.flink.api.common.functions.FlatMapFunction;
@@ -23,9 +6,6 @@ import org.gradoop.common.model.impl.pojo.Vertex;
 
 import java.util.Set;
 
-/**
- * Collects all users and clients given by sets into one dataset.
- */
 public class UserClients
   implements FlatMapFunction<Set<Vertex>, Vertex> {
